# swagger.yml
openapi: 3.0.3
info:
  title: Swagger
  contact:
    email: chris.syst3@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:8080
paths:
  /api/company:
    get:
      tags:
        - empresa
      summary: Obtener lista de empresas
      description: Obtiene una lista de todas las empresas
      responses:
        '200':
          description: Lista de empresas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyRequest'
      x-curl-examples:
        summary: Obtener lista de todas las empresas
        value: |
          curl --location 'http://localhost:8080/api/company'
    put:
      tags:
        - empresa
      summary: documentacion swagger
      description: actualiza una empresa
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyRequest'
    post:
      tags:
        - empresa
      summary: documentacion swagger
      description: crea una empresa
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyRequest'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
  /api/company/{id}:
    get:
      tags:
        - empresa
      summary: Obtener empresa por ID
      description: Obtiene la información de una empresa por su ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la empresa
      responses:
        '200':
          description: Empresa encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyRequest'
        '404':
          description: Empresa no encontrada
      x-curl-examples:
        summary: Obtener empresa con ID 14
        value: |
          curl --location 'http://localhost:8080/api/company/14'
    delete:
      tags:
        - empresa
      summary: Eliminar empresa
      description: Elimina una empresa por su ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la empresa a eliminar
      responses:
        '200':
          description: Empresa eliminada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyRequest'
components:
  schemas:
    CompanyRequest:
      required:
        - id
        - name
        - issuer
        - address
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Christopher
        issuer:
          type: string
          example: test
        address:
          type: string
          example: zona 1